Add linters
