+++ src/components/Navbar.js
@@ -17,13 +17,11 @@ const Navbar = () => {
   return (
     <nav className="navBar">
       <ul>
-        {links.map((link) => {
-          return (
-            <li key={link.id}>
-              <NavLink to={link.path}>{link.text}</NavLink>
-            </li>
-          );
-        })}
+        {links.map((link) => (
+          <li key={link.id}>
+            <NavLink to={link.path}>{link.text}</NavLink>
+          </li>
+        ))}
       </ul>
     </nav>
   );
