+++ src/components/TodoContainer.js
@@ -1,113 +1,95 @@
+/* eslint-disable no-param-reassign */
+
 import { Route, Routes } from 'react-router-dom';
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import { v4 as uuidv4 } from 'uuid';
 import Header from './Header';
 import TodosList from './TodosList';
 import InputTodo from './InputTodo';
-import { v4 as uuidv4 } from 'uuid';
 import About from '../pages/About';
 import NotMatch from '../pages/NotMatch';
 import Navbar from './Navbar';
 import SinglePage from '../pages/SinglePage';
 
-class TodoContainer extends React.Component {
-  state = {
-    todos: [],
-  };
+const TodoContainer = () => {
+  function getInitialTodos() {
+    // getting stored items
+    const temp = localStorage.getItem('todos');
+    const savedTodos = JSON.parse(temp);
+    return savedTodos || [];
+  }
+  const [todos, setTodos] = useState(getInitialTodos());
 
-  handleChange = (id) => {
-    // const todos = this.state.todos.map((todo) => {
-    //   if (todo.id === id) todo.completed = !todo.completed;
-
-    //   return todo;
-    // });
-    // this.setState({ todos });
-    this.setState((prevState) => ({
-      todos: prevState.todos.map((todo) => {
-        if (todo.id === id) {
-          return {
-            ...todo,
-            completed: !todo.completed,
-          };
-        }
-        return todo;
-      }),
+  const handleChange = (id) => {
+    setTodos((prevState) => prevState.todos.map((todo) => {
+      if (todo.id === id) {
+        return {
+          ...todo,
+          completed: !todo.completed,
+        };
+      }
+      return todo;
     }));
   };
 
-  handleDelete = (id) => {
-    this.setState({
-      todos: this.state.todos.filter((todo) => todo.id !== id),
-    });
+  const handleDelete = (id) => {
+    setTodos([...todos.filter((todo) => todo.id !== id)]);
   };
 
-  addTodoItem = (title) => {
+  const addTodoItem = (title) => {
     const newTodo = {
       id: uuidv4(),
-      title: title,
+      title,
       completed: false,
     };
-    this.setState({
-      todos: [...this.state.todos, newTodo],
-    });
+    setTodos([...todos, newTodo]);
   };
 
-  setUpdate = (updatedTitle, id) => {
-    this.setState({
-      todos: this.state.todos.map((todo) => {
+  const setUpdate = (updatedTitle, id) => {
+    setTodos([
+      ...todos.map((todo) => {
         if (todo.id === id) {
           todo.title = updatedTitle;
         }
         return todo;
       }),
-    });
+    ]);
   };
 
-  componentDidMount() {
-    const temp = localStorage.getItem('todos');
-    const loadedTodos = JSON.parse(temp);
-    if (loadedTodos) {
-      this.setState({
-        todos: loadedTodos,
-      });
-    }
-  }
+  useEffect(() => {
+    // storing todos items
+    const temp = JSON.stringify(todos);
+    localStorage.setItem('todos', temp);
+  }, [todos]);
 
-  componentDidUpdate(prevProps, prevState) {
-    if (prevState.todos !== this.state.todos) {
-      const temp = JSON.stringify(this.state.todos);
-      localStorage.setItem('todos', temp);
-    }
-  }
-
-  render() {
-    return (
-      <>
-        <Navbar />
-        <Routes>
-          <Route
-            path="/"
-            element={
-              <div className="container">
-                <div className="inner">
-                  <Header />
-                  <InputTodo addTodoProps={this.addTodoItem} />
-                  <TodosList
-                    todos={this.state.todos}
-                    handleChangeProps={this.handleChange}
-                    handleDeleteProps={this.handleDelete}
-                    setUpdate={this.setUpdate}
-                  />
-                </div>
+  return (
+    <>
+      <Navbar />
+      <Routes>
+        <Route
+          path="/"
+          element={(
+            <div className="container">
+              <div className="inner">
+                <Header />
+                <InputTodo addTodoProps={addTodoItem} />
+                <TodosList
+                  todos={todos}
+                  handleChangeProps={handleChange}
+                  handleDeleteProps={handleDelete}
+                  setUpdate={setUpdate}
+                />
               </div>
-            }
-          />
-          <Route path="/about" element={<About />}>
-            <Route path=":slug" element={<SinglePage />} />
-          </Route>
-          <Route path="*" element={<NotMatch />} />
-        </Routes>
-      </>
-    );
-  }
-}
+            </div>
+          )}
+        />
+        <Route path="/about" element={<About />}>
+          <Route path=":slug" element={<SinglePage />} />
+        </Route>
+        <Route path="*" element={<NotMatch />} />
+      </Routes>
+    </>
+  );
+};
+
 export default TodoContainer;
