+++ src/components/TodoItem.js
@@ -1,73 +1,67 @@
-import React from 'react';
+/* eslint-disable react/destructuring-assignment */
+/* eslint-disable react/prop-types */
+import React, { useState, useEffect } from 'react';
 import styles from './TodoItem.module.scss';
 
-class TodoItem extends React.Component {
-  state = {
-    editing: false,
+const TodoItem = (props) => {
+  const [editing, setEditing] = useState(false);
+
+  const handleEditing = () => {
+    setEditing(true);
   };
 
-  completedStyle = {
+  const handleUpdatedDone = (event) => {
+    if (event.key === 'Enter') {
+      setEditing(false);
+    }
+  };
+
+  const completedStyle = {
     fontStyle: 'italic',
     color: '#595959',
     opacity: 0.4,
     textDecoration: 'line-through',
   };
 
-  handleEditing = () => {
-    this.setState({
-      editing: true,
-    });
-  };
+  const { completed, id, title } = props.todo;
 
-  handleUpdatedDone = (event) => {
-    if (event.key === 'Enter') {
-      this.setState({ editing: false });
-    }
-  };
+  const viewMode = {};
+  const editMode = {};
 
-  componentWillUnmount() {
-    console.log('Cleaning up...');
+  if (editing) {
+    viewMode.display = 'none';
+  } else {
+    editMode.display = 'none';
   }
 
-  render() {
-    const { completed, id, title } = this.props.todo;
+  useEffect(() => () => {}, []);
 
-    let viewMode = {};
-    let editMode = {};
-
-    if (this.state.editing) {
-      viewMode.display = 'none';
-    } else {
-      editMode.display = 'none';
-    }
-
-    return (
-      <li className={styles.item}>
-        <div onDoubleClick={this.handleEditing} style={viewMode}>
-          <input
-            type="checkbox"
-            className={styles.checkbox}
-            checked={completed}
-            onChange={() => this.props.handleChangeProps(id)}
-          />
-          <button onClick={() => this.props.handleDeleteProps(id)}>
-            Delete
-          </button>
-          <span style={completed ? this.completedStyle : null}>{title}</span>
-        </div>
+  return (
+    <li className={styles.item}>
+      <div onDoubleClick={handleEditing} style={viewMode}>
         <input
-          type="text"
-          style={editMode}
-          className={styles.textInput}
-          value={title}
-          onChange={(e) => {
-            this.props.setUpdate(e.target.value, id);
-          }}
-          onKeyDown={this.handleUpdatedDone}
+          type="checkbox"
+          className={styles.checkbox}
+          checked={completed}
+          onChange={() => props.handleChangeProps(id)}
         />
-      </li>
-    );
-  }
-}
+        <button type="button" onClick={() => props.handleDeleteProps(id)}>
+          Delete
+        </button>
+        <span style={completed ? completedStyle : null}>{title}</span>
+      </div>
+      <input
+        type="text"
+        style={editMode}
+        className={styles.textInput}
+        value={title}
+        onChange={(e) => {
+          props.setUpdate(e.target.value, id);
+        }}
+        onKeyDown={handleUpdatedDone}
+      />
+    </li>
+  );
+};
 
 export default TodoItem;
