+++ src/components/TodosList.js
@@ -1,22 +1,20 @@
+/* eslint-disable react/destructuring-assignment */
+/* eslint-disable react/prop-types */
 import React from 'react';
 import TodoItem from './TodoItem';
 
-class TodosList extends React.Component {
-  render() {
-    return (
-      <ul>
-        {this.props.todos.map((todo) => (
-          <TodoItem
-            key={todo.id}
-            todo={todo}
-            handleChangeProps={this.props.handleChangeProps}
-            handleDeleteProps={this.props.handleDeleteProps}
-            setUpdate={this.props.setUpdate}
-          />
-        ))}
-      </ul>
-    );
-  }
-}
+const TodosList = (props) => (
+  <ul>
+    {props.todos.map((todo) => (
+      <TodoItem
+        key={todo.id}
+        todo={todo}
+        handleChangeProps={props.handleChangeProps}
+        handleDeletProps={props.handleDeleteProps}
+        setUpdate={props.setUpdate}
+      />
+    ))}
+  </ul>
+);
 
 export default TodosList;
