+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-//component file
+// component file
 import TodoContainer from './components/TodoContainer';
 
-//stylesheet
+// stylesheet
 import './App.css';
 
 ReactDOM.render(
@@ -13,5 +13,5 @@ ReactDOM.render(
       <TodoContainer />
     </Router>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
