+++ src/pages/NotMatch.js
@@ -1,10 +1,8 @@
 import React from 'react';
 
-const NotMatch = () => {
-  return (
-    <div>
-      <h3>No match for this page</h3>
-    </div>
-  );
-};
+const NotMatch = () => (
+  <div>
+    <h3>No match for this page</h3>
+  </div>
+);
 export default NotMatch;
